+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG for FlatCAM beta
 - added some more logs to the SVG import
 - fixed minor issue with not having the Cancel option on exiting the GCode Editor through the Toolbar button
 - in Gerber Editor fixed the canvas context menu not showing
+- made sure that the Offset fields in the Properties Tab are updated on object: move, origin change, transformations
 
 19.01.2021